Cast enum as unit8

+++ components/balboa_spa/sensor/sensors.cpp
@@ -21,7 +21,7 @@ void BalboaSpaSensors::update() {
             sensor_state_value = spaState->circulation;
             break;
         default:
-             ESP_LOGD("Spa/Sensors/UnknownSensorType", String(sensor_type, 0).c_str());
+             ESP_LOGD("Spa/Sensors/UnknownSensorType", String((uint8_t)sensor_type, 0).c_str());
             // Unknown enum value. Ignore
             return;
     }
